+++ src/cobib/commands/search.py
@@ -53,7 +53,7 @@ import logging
 from typing import List, Type
 
 from rich.console import Console, ConsoleRenderable
-from rich.prompt import PromptBase
+from rich.prompt import PromptBase, PromptType
 from rich.text import Text
 from rich.tree import Tree as RichTree
 from textual.app import App
@@ -94,7 +94,7 @@ class SearchCommand(Command):
         self,
         *args: str,
         console: Console | App[None] | None = None,
-        prompt: Type[PromptBase[str]] | None = None,
+        prompt: Type[PromptBase[PromptType]] | None = None,
     ) -> None:
         super().__init__(*args, console=console, prompt=prompt)